requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
h11==0.8.1
httptools==0.0.13
idna==2.8
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.0.3
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-multipart==0.0.5
PyWavelets==1.0.3
requests==2.21.0
rockethub==0.1.1
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
starlette==0.11.4
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
urllib3==1.24.2
uvicorn==0.7.0
uvloop==0.12.2
websockets==7.0
